[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14116 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -8,6 +8,6 @@ public class Main {
         int A = Integer.parseInt(st.nextToken());
         int B = Integer.parseInt(st.nextToken());
 
-        System.out.println((A + B) * (A - B));
+        System.out.println((long) (A + B) * (A - B));
     }
 }
